+++ task_kanghyun_do_4/jsfile/post-body-fetch.js
@@ -1,0 +1,13 @@
+fetch("https://jsonplaceholder.typicode.com/posts")
+    .then((response) => response.json())
+    .then((posts) => {
+        posts.forEach((post) => {
+            const { body } = post;
+
+            postInfo(body);
+        });
+    });
+
+function postInfo(body) {
+    console.log(body);
+}
