+++ task_kanghyun_do_4/jsfile/post-title-fetch.js
@@ -1,0 +1,13 @@
+fetch("https://jsonplaceholder.typicode.com/posts")
+    .then((response) => response.json())
+    .then((posts) => {
+        posts.forEach((post) => {
+            const { title } = post;
+
+            postInfo(title);
+        });
+    });
+
+function postInfo(title) {
+    console.log(title);
+}
